small adjustments
