+++ packages/gatsby-theme-bodiless/gatsby-config.js
@@ -12,7 +12,6 @@
  * limitations under the License.
  */
 
-const fs = require('fs');
 const path = require('path');
 
 require('dotenv').config({
@@ -128,65 +127,9 @@ if (process.env.ROBOTSTXT_ENABLED !== '0') {
 /**
  * css compilation and purging.
 */
+const getbuildCSSPlugins = require('./build-css');
 
-const tailwindThemeEnabled = (process.env.BODILESS_TAILWIND_THEME_ENABLED || '1') === '1';
-const getDirectories = source => {
-  if (!fs.existsSync(source)) {
-    return [];
-  }
-  return fs.readdirSync(source, { withFileTypes: true })
-    .filter(dirent => dirent.isDirectory() || dirent.isSymbolicLink())
-    .map(dirent => dirent.name);
-};
-
-const globPattern = '**/!(*.d).{ts,js,jsx,tsx}';
-const bodilessPackagesBasePath = path.resolve('./node_modules/@bodiless');
-const bodilessFilesPaths = getDirectories(bodilessPackagesBasePath)
-  .map(pkg => path.resolve(bodilessPackagesBasePath, pkg))
-  .map(dir => path.resolve(dir, globPattern));
-
-const getTailwindConfig = () => {
-  const processDirFile = path.resolve(process.cwd(), 'tailwind.config.js');
-  if (fs.existsSync(processDirFile)) {
-    return processDirFile;
-  }
-  const siteDirFile = path.resolve(__dirname, 'tailwind.config.js');
-  if (fs.existsSync(siteDirFile)) {
-    return siteDirFile;
-  }
-  return false;
-};
-const tailWindConfigFile = getTailwindConfig();
-
-if (process.env.BODILESS_PURGE_CSS_ENABLED !== '0') {
-  if (tailwindThemeEnabled && tailWindConfigFile) {
-    plugins.push({
-      resolve: 'gatsby-plugin-postcss',
-      options: {
-        postCssPlugins: [
-          // eslint-disable-next-line global-require
-          require('tailwindcss')(tailWindConfigFile),
-        ],
-      },
-    });
-  }
-  plugins.push({
-    resolve: 'gatsby-plugin-purgecss',
-    options: {
-      tailwind: true,
-      purgeOnly: [
-        'src/css/style.css',
-        'packages/bodiless-ui/lib/bodiless.index.css',
-        'node_modules/@bodiless/ui/lib/bodiless.index.css',
-      ],
-      content: [
-        path.join(process.cwd(), 'src/**/!(*.d).{ts,js,jsx,tsx}'),
-        path.join(process.cwd(), 'node_modules/@bodiless/**/!(*.d).{ts,js,jsx,tsx}'),
-        ...bodilessFilesPaths,
-      ],
-    },
-  });
-}
+plugins.push(...getbuildCSSPlugins());
 
 module.exports = {
   siteMetadata: {
